Bug 427399 - Provide an Orion Node shell based on tty.js
Moved support to allow shells to exist simultaneously
