+++ modules/orionode/index.js
@@ -17,8 +17,7 @@ var connect = require('connect'),
     orionWorkspace = require('./lib/workspace'),
     orionNodeStatic = require('./lib/orionode_static'),
     orionStatic = require('./lib/orion_static'),
-    term = require('term.js'),
-    pty = require('pty.js');
+    term = require('term.js');
 
 var LIBS = path.normalize(path.join(__dirname, 'lib/')),
     ORION_CLIENT = path.normalize(path.join(__dirname, '../../'));
@@ -57,55 +56,11 @@ function startServer(options) {
 				appContext: appContext,
 				root: '/node'
 			}));
-		app.appContext = appContext;
+
 		// Socket server
-		//var io = socketio.listen(app, { 'log level': 1 });
+    var io = socketio.listen(app, { 'log level': 1 });
 
-    // Terminal Socket
-    var termio = socketio.listen(app, { 'log level': 1 });
-
-    termio.sockets.on('connection', function(sock) {
-
-      var buff = [];
-      // Open Terminal Connection
-      var terminal = pty.fork(process.env.SHELL || 'sh', [], {
-        name: require('fs').existsSync('/usr/share/terminfo/x/xterm-256color')
-        ? 'xterm-256color'
-        : 'xterm',
-               cols: 80,
-               rows: 24,
-               cwd: process.env.HOME
-      });
-
-      terminal.on('data', function(data) {
-        return !sock
-        ? buff.push(data)
-        : sock.emit('data', data);
-      });
-
-      console.log(''
-        + 'Created shell h pty master/slave'
-        + ' pair (master: %d, pid: %d)',
-        terminal.fd, terminal.pid);
-
-      // Set up communication paths
-      sock.on('data', function(data) {
-        terminal.write(data);
-      });
-
-      sock.on('disconnect', function() {
-        terminal.destroy()
-        termsocket = null;
-      });
-
-      while (buff.length) {
-        sock.emit('data', buff.shift());
-      }
-    });
-
-
-	  //appSocket.install({io: io, appContext: appContext});
-		appSocket.install({io: termio, appContext: appContext});
+		appSocket.install({io: io, appContext: appContext});
 		app.on('error', handleError);
 		return app;
 	} catch (e) {
